+++ PSFilterShim/PSApi/LoadPsFilter.cs
@@ -4947,11 +4947,11 @@ namespace PSFilterLoad.PSApi
 			descriptorParameters->writeDescriptorProcs = writeDescriptorPtr;
 			if (!isRepeatEffect)
 			{
-				descriptorParameters->recordInfo = (short)RecordInfo.plugInDialogOptional;
+				descriptorParameters->recordInfo = RecordInfo.plugInDialogOptional;
 			}
 			else
 			{
-				descriptorParameters->recordInfo = (short)RecordInfo.plugInDialogNone;
+				descriptorParameters->recordInfo = RecordInfo.plugInDialogNone;
 			}
 
 
@@ -4961,16 +4961,16 @@ namespace PSFilterLoad.PSApi
 
 				if (!isRepeatEffect)
 				{
-					descriptorParameters->playInfo = (short)PlayInfo.plugInDialogDisplay;
+					descriptorParameters->playInfo = PlayInfo.plugInDialogDisplay;
 				}
 				else
 				{
-					descriptorParameters->playInfo = (short)PlayInfo.plugInDialogDontDisplay;
+					descriptorParameters->playInfo = PlayInfo.plugInDialogDontDisplay;
 				}
 			}
 			else
 			{
-				descriptorParameters->playInfo = (short)PlayInfo.plugInDialogDisplay;
+				descriptorParameters->playInfo = PlayInfo.plugInDialogDisplay;
 			}
 
 			if (useChannelPorts)
